#101: Remove experimental flag
- We are sure that this path is not invoked by current versions of mybatis,
and if it was, iut would have failed, it is thus safe to remove the flag
