+++ src/main/java/org/apache/ibatis/builder/xml/XMLConfigBuilder.java
@@ -292,8 +292,6 @@ public class XMLConfigBuilder extends BaseBuilder {
         booleanValueOf(props.getProperty("argNameBasedConstructorAutoMapping"), false));
     configuration.setDefaultSqlProviderType(resolveClass(props.getProperty("defaultSqlProviderType")));
     configuration.setNullableOnForEach(booleanValueOf(props.getProperty("nullableOnForEach"), false));
-    configuration.setExperimentalConstructorCollectionMapping(
-        booleanValueOf(props.getProperty("experimentalConstructorCollectionMapping"), false));
   }
 
   private void environmentsElement(XNode context) throws Exception {
