+++ src/main/java/org/apache/ibatis/session/Configuration.java
@@ -117,7 +117,6 @@ public class Configuration {
   protected boolean shrinkWhitespacesInSql;
   protected boolean nullableOnForEach;
   protected boolean argNameBasedConstructorAutoMapping;
-  protected boolean experimentalConstructorCollectionMapping;
 
   protected String logPrefix;
   protected Class<? extends Log> logImpl;
@@ -368,14 +367,6 @@ public class Configuration {
 
   public boolean isSafeRowBoundsEnabled() {
     return safeRowBoundsEnabled;
-  }
-
-  public void setExperimentalConstructorCollectionMapping(boolean experimentalConstructorCollectionMapping) {
-    this.experimentalConstructorCollectionMapping = experimentalConstructorCollectionMapping;
-  }
-
-  public boolean isExperimentalConstructorCollectionMappingEnabled() {
-    return experimentalConstructorCollectionMapping;
   }
 
   public void setSafeRowBoundsEnabled(boolean safeRowBoundsEnabled) {
