+++ src/site/markdown/configuration.md
@@ -137,7 +137,6 @@ These are extremely important tweaks that modify the way that MyBatis behaves at
 | defaultSqlProviderType                   | Specifies an sql provider class that holds provider method (Since 3.5.6). This class apply to the `type`(or `value`) attribute on sql provider annotation(e.g. `@SelectProvider`), when these attribute was omitted.                                                                                                                                                                                                                             | A type alias or fully qualified class name                                                                                                 | Not set                                               |
 | nullableOnForEach                        | Specifies the default value of 'nullable' attribute on 'foreach' tag. (Since 3.5.9)                                                                                                                                                                                                                                                                                                                                                              | true &#124; false                                                                                                                          | false                                                 |
 | argNameBasedConstructorAutoMapping       | When applying constructor auto-mapping, argument name is used to search the column to map instead of relying on the column order. (Since 3.5.10)                                                                                                                                                                                                                                                                                                 | true &#124; false                                                                                                                          | false                                                 |
-| experimentalConstructorCollectionMapping | When applying constructor mapping and any nested result mappings containing collections are found, they will now be filled (Since 3.6.0)                                                                                                                                                                                                                                                                                                         | true &#124; false                                                                                                                          | false                                                 |
 
 An example of the settings element fully configured is as follows:
