+++ src/test/resources/org/apache/ibatis/submitted/collection_injection/mybatis_config.xml
@@ -21,9 +21,6 @@
     "https://mybatis.org/dtd/mybatis-3-config.dtd">
 
 <configuration>
-    <settings>
-        <setting name="experimentalConstructorCollectionMapping" value="true"/>
-    </settings>
 
     <typeAliases>
         <typeAlias type="org.apache.ibatis.submitted.collection_injection.property.House" alias="House"/>
